refactor: create createElement.ts
